🚚 Move replaceAllCastedVar to variableCastingHelpers
